split mono clock into separate files
