+++ x/cpp/telem/BUILD.bazel
@@ -12,6 +12,7 @@ cc_library(
         "clock_skew.h",
         "frame.h",
         "json.gen.h",
+        "mono_clock.h",
         "proto.gen.h",
         "series.h",
         "telem.h",
@@ -45,6 +46,7 @@ cc_test(
     srcs = [
         "clock_skew_test.cpp",
         "frame_test.cpp",
+        "mono_clock_test.cpp",
         "series_test.cpp",
         "telem_test.cpp",
     ],
