+++ x/go/telem/time_stamp.go
@@ -44,31 +44,6 @@ func (ts TimeStamp) MarshalJSON() ([]byte, error) {
 // Now returns the current time as a TimeStamp.
 func Now() TimeStamp { return NewTimeStamp(time.Now()) }
 
-// MonoClock provides monotonically increasing timestamps. On platforms with
-// coarse clock resolution (e.g. Windows), consecutive Now() calls can return
-// the same value. MonoClock guarantees each call to Now() returns a strictly
-// greater timestamp than the previous one by bumping by 1 nanosecond when
-// necessary. A custom time source can be provided for testing; if nil, the
-// package-level Now function is used.
-type MonoClock struct {
-	last   TimeStamp
-	Source func() TimeStamp
-}
-
-// Now returns a timestamp that is strictly greater than any previous call.
-func (c *MonoClock) Now() TimeStamp {
-	source := c.Source
-	if source == nil {
-		source = Now
-	}
-	now := source()
-	if now <= c.last {
-		now = c.last + 1
-	}
-	c.last = now
-	return now
-}
-
 // NewTimeStamp creates a new TimeStamp from a time.Time.
 func NewTimeStamp(t time.Time) TimeStamp { return TimeStamp(t.UnixNano()) }
